:truck: Move DeviceHandler to /devices/
