+++ core/narc_engine/include/buffers/Buffer.h
@@ -2,7 +2,7 @@
 
 #include <vulkan/vulkan.h>
 
-#include "include/DeviceHandler.h"
+#include "include/devices/DeviceHandler.h"
 
 namespace narc_engine
 {
