+++ core/narc_engine/src/SwapChain.cpp
@@ -1,7 +1,7 @@
 #include "include/SwapChain.h"
 
 #include "include/Engine.h"
-#include "include/DeviceHandler.h"
+#include "include/devices/DeviceHandler.h"
 
 namespace narc_engine
 {
